fix(bananass-build): correct `clean` option behavior (#29)
This pull request includes changes to the
`packages/bananass/src/commands/bananass-build/webpack.js` file to
improve the handling of the output directory cleaning process during the
build.

Improvements to output directory cleaning:

* Added `rmSync` from `node:fs` to manually clean the output directory
before emitting files.
* Commented out the `webpackConfigs.output.clean` option to avoid using
it, as it could result in only one file in the output directory.
* Implemented a manual cleaning process using `rmSync` to prevent race
conditions and ensure the output directory is cleaned only once.