chore: create `.gitattributes` for consistent EOL handling (#31)
This pull request includes changes to ensure consistent end-of-line
(EOL) settings across all text files and to synchronize the
`.gitattributes` file with the project configurations.

EOL consistency:

*
[`.gitattributes`](diffhunk://#diff-618cd5b83d62060ba3d027e314a21ceaf75d36067ff820db126642944145393eR1-R4):
Added settings to ensure consistent EOL (LF) for all text files.

Project configuration synchronization:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R52-R53):
Included the `.gitattributes` file in the list of files to be
synchronized with the project configurations.