refactor(bananass-utils-console): delete `utils` directory and replace it with dependencies (#35)
This pull request includes several changes to the
`bananass-utils-console` package, focusing on dependency updates and
code simplification. The most important changes include updating
dependencies in the `package.json` file, refactoring the code to use
external libraries for Unicode support and interactivity checks, and
removing redundant utility functions.

Dependency updates:

*
[`packages/bananass-utils-console/package.json`](diffhunk://#diff-14dc0ab8d918f740c0f0d225fcfbbfdbf6620ab25f1e833d884d3878b35a9a0eL61-R63):
Added `is-interactive` and `is-unicode-supported` dependencies.

Code refactoring:

*
[`packages/bananass-utils-console/src/icons/index.js`](diffhunk://#diff-c14edb8e4ab6ed0399ee7474a0b6b2ed5a227761c5e7ea23877f29cb2575324cL12-R25):
Updated imports to use `is-unicode-supported` directly and modified icon
definitions to call the function.
*
[`packages/bananass-utils-console/src/spinner.js`](diffhunk://#diff-a57af170b8c724600897a5102de4eb2c0a4360d0e34086c8867da06cfc52d6ebR14-R16):
Imported `is-interactive` and updated the `Spinner` class to use the new
dependency.
[[1]](diffhunk://#diff-a57af170b8c724600897a5102de4eb2c0a4360d0e34086c8867da06cfc52d6ebR14-R16)
[[2]](diffhunk://#diff-a57af170b8c724600897a5102de4eb2c0a4360d0e34086c8867da06cfc52d6ebL53-R54)

Code cleanup:

*
[`packages/bananass-utils-console/src/utils/index.js`](diffhunk://#diff-2a54a535d0cd98a34be5b433ded60c52f38a2892a4d3365bcac7a7fadba04c29L1-L36):
Removed the custom implementations of `isUnicodeSupported` and
`isInteractive` functions.