test(bananass-utils-console): add tests for `spinner` functionality and update dependencies (#36)
This pull request introduces a new test file for the `spinner.js` module
and adds a new development dependency to the `bananass-utils-console`
package. The primary focus is on enhancing the test coverage for the
`spinner.js` functionality.

### Enhancements to test coverage:

*
[`packages/bananass-utils-console/src/spinner.test.js`](diffhunk://#diff-c07c0032e447775e33efc82ed90e5c25c75d2e9437b5acb548190f9008e8874aR1-R156):
Added comprehensive tests for `spinner.js`, including various scenarios
such as starting, stopping, and changing the spinner text, as well as
handling different spinner states like success, error, warning, and
info.

### Dependency updates:

*
[`packages/bananass-utils-console/package.json`](diffhunk://#diff-14dc0ab8d918f740c0f0d225fcfbbfdbf6620ab25f1e833d884d3878b35a9a0eR66):
Added `get-stream` as a new development dependency to facilitate stream
handling in tests.