chore: add `website` workspace and initialize `package.json` (#38)
This pull request includes changes to support the addition of a new
`website` workspace. The most important changes include updating the
`workspaces` configuration in `package.json` and adding a `package.json`
file for the `website` workspace.

Workspace configuration:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L6-R7):
Added the `website` directory to the list of workspaces.

New workspace setup:

*
[`website/package.json`](diffhunk://#diff-fae242fbf77a8a9d52625664bc36ea12316586f8a716b41137d897a2b7e3df76R1-R8):
Created a new `package.json` file for the `website` workspace with basic
configuration including a `start` script.