chore(examples-solutions-readline-esm): add examples for ESM solutions with readline (#43)
This pull request introduces a new example solution for the Baekjoon
problem 1000 using ESM (ECMAScript Modules) in Node.js. The changes
include the addition of a `package.json` file and the implementation of
the solution in a new JavaScript file.

### New Example Solution for Baekjoon Problem 1000:

*
[`examples/solutions-readline-esm/package.json`](diffhunk://#diff-dfaffaa92838021bc82c5f59c0831a2254b46a25409b25fffcc1f03a100cefa9R1-R9):
Added a new `package.json` file to define the project as a module and
include a script for running the solution.

*
[`examples/solutions-readline-esm/src/1000.js`](diffhunk://#diff-595afe0897143ed9aae3e84be5c7a0649e33a6f2de08c405c87914e9033f6d61R1-R50):
Implemented the solution for Baekjoon problem 1000 using ESM syntax,
including imports from Node.js built-in modules and handling user input
via the `readline` module.