refactor(examples-solutions-bananass): refactor functions and update dependency (#45)
This pull request includes several changes to the
`examples/solutions-bananass` project, primarily focused on refactoring
functions and updating dependencies. The most important changes are
listed below:

### Refactoring Functions:

*
[`examples/solutions-bananass/bananass/1000.js`](diffhunk://#diff-cee990384708d6ee19026603ffb858de9ae8dc486da42ebb853c297bb946ac87L3-R12):
Refactored the function to use a named `solution` function instead of an
anonymous function and updated the export statement.
*
[`examples/solutions-bananass/bananass/1001.js`](diffhunk://#diff-9462768f82f8319792e64b734fd98823ce8baf518674d12639966ae6f6f01008L3-R12):
Refactored the function to use a named `solution` function instead of an
anonymous function and updated the export statement.
*
[`examples/solutions-bananass/bananass/29751.js`](diffhunk://#diff-cb36472122717be8cb857d99059e4bd3b6a4e66a3c7872c7100b9ce65e35d6e8L3-R14):
Refactored the function to use a named `solution` function instead of an
anonymous function and updated the export statement.

### Dependency Updates:

*
[`examples/solutions-bananass/package.json`](diffhunk://#diff-9626239dd77dbfa60c8cce8be28c20c0b09c2e9a08489d202fdf01300d3987cbL8-R8):
Changed `dependencies` to `devDependencies` for the `bananass` package.
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L23-R24):
Added a new script `examples:solutions-bananass:build` to run the
`bananass:build` script for the `examples/solutions-bananass` workspace.