refactor(prettier-config-bananass): add peerDependencies, prepublish script, type support (#60)
This pull request includes changes to the `prettier-config-bananass`
package to update dependencies and improve documentation. The most
important changes include adding a `prepublishOnly` script, specifying
`peerDependencies`, and enhancing module documentation.

Updates to dependencies and scripts:

*
[`packages/prettier-config-bananass/package.json`](diffhunk://#diff-0fdaf27a6fc3a9c98ce068e2f60dd6fa1d2a82cfc232695581bcae3256c32d76L32-R38):
Added a `prepublishOnly` script to run the build process before
publishing, and specified `prettier` as a peer dependency.

Enhancements to documentation:

*
[`packages/prettier-config-bananass/src/index.js`](diffhunk://#diff-5704c175b00dfba9748074a44221d8bdeafc9a341ed6c6107bb94b85ee8394f5L5-R12):
Added a link to the Prettier options documentation in the module
overview comment and included a type annotation for the Prettier
configuration.