chore(dependabot): update configuration to include groups, assignees, and reviewers (#71)
This pull request includes updates to the `.github/dependabot.yml` file
to improve dependency management and assignment. The changes introduce
new assignees and reviewers, organize dependencies into groups, and
adjust the versioning strategy.

Dependency management improvements:

* Added `assignees` and `reviewers` fields to ensure that 'lumirlumir'
is assigned to and reviews the dependency updates for both npm and
GitHub Actions.
* Organized npm dependencies into groups (`bananass`, `babel`, `react`,
and `next`) to better manage related dependencies.
* Changed the `versioning-strategy` to 'increase' for npm dependencies
to ensure a consistent update strategy.