chore: update ESLint configuration to use `eslint-config-bananass` and adjust ignored paths (#75)
This pull request includes updates to configuration files and
dependencies to improve code consistency and maintainability. The most
important changes involve updating the ESLint configuration and
synchronizing project files.

Configuration updates:

*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1R72-R73):
Added `eslint.config.js` to the list of files to be synchronized.

Dependency updates:

*
[`eslint.config.js`](diffhunk://#diff-a32a0887ed9d1d707bbb3b845b7df7fd40e673c47e7b60a3ebd896b68d3b8839L1-R5):
Changed the ESLint configuration to use `eslint-config-bananass` instead
of `eslint-config-bananass-react` and updated the list of ignored
directories.
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L39-R39):
Updated the dependency from `eslint-config-bananass-react` to
`eslint-config-bananass`.