chore: restructure website directory and update package configurations for Docusaurus (#86)
This pull request includes changes to the `package.json` file to refine
the project structure and improve the development workflow. The most
important changes include updating the file paths for the `website`
directory and modifying the development script to target the
`docusaurus` subdirectory.

Changes to project structure:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L13-R19):
Updated the file paths to include all subdirectories under `website`
instead of just the root directory.

Improvements to development workflow:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L13-R19):
Modified the `dev` script to start the development server specifically
for the `website/docusaurus` subdirectory.