feat(website-eslint-config-bananass): introduce `website-eslint-config-bananass` (#87)
This pull request includes changes to the `package.json` and the
addition of a new ESLint configuration for the
`website/eslint-config-bananass` package. The most important changes
include updating the build and development scripts, adding a new ESLint
configuration file, and creating a new package for the ESLint
configuration.

Updates to build and development scripts:

*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L18-R24):
Updated the build and development scripts to include specific commands
for `website/docusaurus` and `website/eslint-config-bananass`.

Addition of ESLint configuration:

*
[`website/eslint-config-bananass/config.js`](diffhunk://#diff-ebfefba036261a2441cae62182d67a2a5091aaa256579f3aa66dafdd1d6dda07R1-R3):
Added a new configuration file that requires and exports the recommended
ESLint configuration from `eslint-config-bananass`.
*
[`website/eslint-config-bananass/package.json`](diffhunk://#diff-afaa8e665e360987b3912648ec4a96561c751cff93c32139218d14e6d97dd3b4R1-R13):
Created a new package for the ESLint configuration with scripts for
development and build, and added `eslint-config-bananass` and
`@eslint/config-inspector` as development dependencies.