ci(*): create `pull-request.yml` and update `labeler` logics (#126)
This pull request includes several changes to GitHub Actions and
configuration files to improve the automation and labeling process. The
most important changes include the removal of the labeler workflow and
configuration, updates to the release configuration, and the addition of
a new pull request workflow.

### Removal of Labeler Workflow and Configuration:
*
[`.github/labeler.yml`](diffhunk://#diff-a22c263686553013feaeb0677d26eeb0b8778a756c4311c1fce13384258026aaL1-L52):
Removed the labeler configuration file.
*
[`.github/workflows/labeler.yml`](diffhunk://#diff-09b72f3c9a3e4f00ab00cd7000b302db25f056075d8895bd91b3654d6e7e956bL1-L20):
Removed the labeler workflow.
*
[`.github/sync-client.yml`](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L14-R19):
Removed references to the labeler configuration file.
[[1]](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L14-R19)
[[2]](diffhunk://#diff-93bc202766315b6269beef308a6ad30ed3e86938ddbfa31b49e030f2263695f1L33-L34)

### Updates to Release Configuration:
*
[`.github/release.yml`](diffhunk://#diff-409fea45635f464aa0592700a92cf483be2f5b21b60f8f1b383756067ee79213L6-R66):
Updated labels and categories for the changelog, including the addition
of new categories and exclusion of certain labels.

### Addition of New Pull Request Workflow:
*
[`.github/workflows/pull-request.yml`](diffhunk://#diff-b71166ed0f585913318ed46933ff9b12901e211de3ac88c40de03f0a944c0ae0R1-R284):
Added a new workflow to lint pull request titles and apply labels based
on the Conventional Commits specification.