ci(*): update `pull-request.yml` concurrency group and update `sync-client.yml` (#134)
This pull request includes updates to the GitHub Actions workflows to
improve concurrency handling and formatting of commit messages and pull
request bodies.

Improvements to concurrency handling:

*
[`.github/workflows/pull-request.yml`](diffhunk://#diff-b71166ed0f585913318ed46933ff9b12901e211de3ac88c40de03f0a944c0ae0L8-R8):
Updated the `group` key in the `concurrency` section to use the pull
request number instead of the reference, ensuring unique groups for each
pull request.

Formatting updates:

*
[`.github/workflows/sync-client.yml`](diffhunk://#diff-254f50149bdf646bb0bb19197c9fa73d054fdb550909072ed57e0b5bcc39a197L25-R26):
Modified the `COMMIT_PREFIX` to use a more descriptive format and
updated the `PR_BODY` to improve readability and include an emoji.