feat(bananass): add webpack banner plugin with generated file information (#139)
This pull request includes changes to the `packages/bananass` module to
enhance the build process and improve code documentation. The most
important changes include the addition of a Webpack banner plugin, the
introduction of a new constant for the Webpack banner, and the
reorganization of exports in the `constants.js` file.

### Enhancements to the build process:

*
[`packages/bananass/src/commands/bananass-build/webpack.js`](diffhunk://#diff-0ce95c8c9e2a9153c60042f1f9db2f62324e019f7980eb2e8ee8a5001fd6e34bR127-R131):
Added the Webpack BannerPlugin to include a banner in the generated
files.

### Improvements to code documentation and organization:

*
[`packages/bananass/src/core/constants.js`](diffhunk://#diff-97d6046c52dcaa9f59ecd0bc5a1ed48d3a91d2c8159901b86ba0b4932b532886R37-R63):
Introduced a new constant `WEBPACK_BANNER` for the Webpack banner
content.
*
[`packages/bananass/src/core/constants.js`](diffhunk://#diff-97d6046c52dcaa9f59ecd0bc5a1ed48d3a91d2c8159901b86ba0b4932b532886L20-R28):
Reorganized export comments for better clarity and added type
annotations for string constants.

### Import adjustments:

*
[`packages/bananass/src/commands/bananass-build/webpack.js`](diffhunk://#diff-0ce95c8c9e2a9153c60042f1f9db2f62324e019f7980eb2e8ee8a5001fd6e34bR21):
Added `WEBPACK_BANNER` to the import statements.