refactor(bananass): delete unused `find-existing-path.js` (#141)
This pull request removes the `find-existing-path.js` file and its
references from the `packages/bananass` module. The most important
changes include the deletion of the `findExistingPath` function and its
corresponding import and export statements.

File deletions and updates:

*
[`packages/bananass/src/core/fs/find-existing-path.js`](diffhunk://#diff-dddfa32e4a1118bb8c1a028bf373f938ac3671bfeabf2518f9ff425f69ed4c5fL1-L58):
Removed the entire file, which included the `findExistingPath` function
and its helper functions.
*
[`packages/bananass/src/core/fs/index.js`](diffhunk://#diff-ba3524c4dcdd31e7c10fe73cd493769dffe36dea9b87e5b434f711309597eb2bL2-R3):
Removed the import and export statements for the `findExistingPath`
function.