feat(bananass): create new command `open` (#156)
This pull request includes significant changes to the `bananass`
package, primarily focusing on enhancing the functionality of the CLI by
adding support for opening Baekjoon problem numbers in a browser.
Additionally, it introduces new structures and updates existing
descriptions and constants.

### CLI Enhancements:
*
[`packages/bananass/src/cli/bananass-open.js`](diffhunk://#diff-9d4a836ada9ae83082ec770982e73c8cd1399ef1153f4a61e894471101f6c99bL9-R29):
Replaced the `warning` import with `logger` and added imports for new
commands, configurations, descriptions, arguments, and options. Updated
the `bananassOpen` function to include new options and actions for
opening problems in a browser.
[[1]](diffhunk://#diff-9d4a836ada9ae83082ec770982e73c8cd1399ef1153f4a61e894471101f6c99bL9-R29)
[[2]](diffhunk://#diff-9d4a836ada9ae83082ec770982e73c8cd1399ef1153f4a61e894471101f6c99bL29-R79)

### Command Additions:
*
[`packages/bananass/src/commands/bananass-open/open.js`](diffhunk://#diff-6c43147143109854fa17abf461dfdde7e862fc194454f5869fae5dc60631a682R1-R87):
Added a new command to asynchronously open given Baekjoon problem
numbers in a browser, including runtime validation and CLI animations.
*
[`packages/bananass/src/commands/bananass-open/index.js`](diffhunk://#diff-6cfc0ed8b572f63c1df81e4fbd87914243c3ce1c003fb440a3be265c257e43eaR1-R3):
Exported the new `open` command.
*
[`packages/bananass/src/commands/index.js`](diffhunk://#diff-c96aa353b75dcf3f251899c2993ade50ecacb5517ebbf35188abd138b982e1cdR11-R17):
Included the new `open` command in the exported commands.

### Structural Updates:
*
[`packages/bananass/src/core/structs/problem/problem.js`](diffhunk://#diff-b9e69c5b70280e5781d72c0dd86060f4b82d9dbd632d1648928d099e9eb55544R1-R40):
Introduced a new `Problem` type struct to validate Baekjoon problem
numbers.
*
[`packages/bananass/src/core/structs/problems/problems.js`](diffhunk://#diff-acf357c84604c79c20cdbcf1493ab589dea49efbe1f4eabdaaf76566fe3b0c17L30-R35):
Updated the `Problems` struct to use the new `Problem` type for
validation.

### Descriptions and Constants:
*
[`packages/bananass/src/core/cli/descriptions.js`](diffhunk://#diff-4779ca37913f14f5aec15ce530617d866fc03de73d6a32a61264fbc8bdef8a77L18-R20):
Updated descriptions to reflect the new functionality of opening
problems in a browser.
*
[`packages/bananass/src/core/constants.js`](diffhunk://#diff-97d6046c52dcaa9f59ecd0bc5a1ed48d3a91d2c8159901b86ba0b4932b532886R52-R55):
Added new constants for Baekjoon URLs and updated existing ones.

These changes collectively enhance the functionality and robustness of
the `bananass` CLI tool, making it more versatile and user-friendly.