feat(bananass): create new `repo` command (#161)
This pull request introduces a new `repo` command to the `bananass` CLI
tool. The most important changes include adding the new command,
updating the imports and exports, and implementing the functionality to
open the GitHub repository in a browser.

New `repo` command:

*
[`packages/bananass/src/cli/bananass-repo.js`](diffhunk://#diff-18cad518a6aa21886b7a585bd6846ed203fcbb98516a9191ed3c29e82e30869cR1-R77):
Added a new file to define the `repo` command, including options and the
action to open the GitHub repository in a browser.
*
[`packages/bananass/src/commands/bananass-repo/repo.js`](diffhunk://#diff-561777915366eed77e0e117f652a202ef5a969d05ae4b8b4f4d3291f03102088R1-R78):
Implemented the function to asynchronously open the GitHub repository in
a browser, with runtime validation and error handling.
*
[`packages/bananass/src/commands/bananass-repo/repo.test.js`](diffhunk://#diff-fb3d79f6174c4b97bc708483d7bc4f71dcf9e98b65ef3e1d8b6ba0f02a653080R1):
Added a placeholder for future tests of the `repo` command.

Updates to imports and exports:

*
[`packages/bananass/src/cli.js`](diffhunk://#diff-6544fc29ddb1c7332552c44ae9262c59f8067412edbd9d71bbfe3edd7472813dR24):
Added the `repo` command to the list of imports and invoked it within
the CLI program.
[[1]](diffhunk://#diff-6544fc29ddb1c7332552c44ae9262c59f8067412edbd9d71bbfe3edd7472813dR24)
[[2]](diffhunk://#diff-6544fc29ddb1c7332552c44ae9262c59f8067412edbd9d71bbfe3edd7472813dR45)
*
[`packages/bananass/src/cli/index.js`](diffhunk://#diff-c9e0cfc980e6dc61abd041c44f2a2caafdfff35f9554d653ff592f72b8d30aceR12):
Updated imports and exports to include the new `repo` command.
[[1]](diffhunk://#diff-c9e0cfc980e6dc61abd041c44f2a2caafdfff35f9554d653ff592f72b8d30aceR12)
[[2]](diffhunk://#diff-c9e0cfc980e6dc61abd041c44f2a2caafdfff35f9554d653ff592f72b8d30aceR29)
*
[`packages/bananass/src/commands/index.js`](diffhunk://#diff-c96aa353b75dcf3f251899c2993ade50ecacb5517ebbf35188abd138b982e1cdR12-R18):
Added the `repo` command to the exports.
*
[`packages/bananass/src/core/cli/descriptions.js`](diffhunk://#diff-4779ca37913f14f5aec15ce530617d866fc03de73d6a32a61264fbc8bdef8a77R21-R22):
Added a description for the `repo` command.

These changes collectively introduce and integrate the new `repo`
command into the `bananass` CLI tool, allowing users to open the GitHub
repository directly from the command line.