feat(bananass): create new `bug` command (#166)
This pull request introduces a new `bug` command to the `bananass` CLI,
allowing users to open GitHub issues in a browser. The changes include
adding the necessary imports, creating the command implementation, and
updating relevant files to integrate the new command.

Key changes include:

### New Command Implementation:
*
[`packages/bananass/src/cli/bananass-bug.js`](diffhunk://#diff-196f36c2928080b1421ad5f1091b6715ba428f781678296dc43ee5fd1b7f4bd4R1-R80):
Added the implementation of the `bug` command, including importing
required modules, defining the command, and handling its options and
actions.

### Integration of the New Command:
*
[`packages/bananass/src/cli.js`](diffhunk://#diff-6544fc29ddb1c7332552c44ae9262c59f8067412edbd9d71bbfe3edd7472813dR15):
Added the `bug` command to the list of commands imported and registered
with the CLI program.
[[1]](diffhunk://#diff-6544fc29ddb1c7332552c44ae9262c59f8067412edbd9d71bbfe3edd7472813dR15)
[[2]](diffhunk://#diff-6544fc29ddb1c7332552c44ae9262c59f8067412edbd9d71bbfe3edd7472813dR38)
*
[`packages/bananass/src/cli/index.js`](diffhunk://#diff-c9e0cfc980e6dc61abd041c44f2a2caafdfff35f9554d653ff592f72b8d30aceR3):
Imported the `bug` command module and included it in the exported
commands.
[[1]](diffhunk://#diff-c9e0cfc980e6dc61abd041c44f2a2caafdfff35f9554d653ff592f72b8d30aceR3)
[[2]](diffhunk://#diff-c9e0cfc980e6dc61abd041c44f2a2caafdfff35f9554d653ff592f72b8d30aceR22)

### Command Functionality:
*
[`packages/bananass/src/commands/bananass-bug/bug.js`](diffhunk://#diff-c49c2ac0f48f162fb6dd63de6eaaf35f7fc82f0372fc582a3332539634eb576bR1-R78):
Implemented the functionality to open GitHub issues in a browser,
including handling browser options and logging.
*
[`packages/bananass/src/commands/bananass-bug/index.js`](diffhunk://#diff-db9d4d381fb9c38bc27a6388b1c1c7b6881ee033fd0dcbaab68cb9e44abe8057R1-R3):
Added an index file to export the `bug` command.

### Command Description:
*
[`packages/bananass/src/core/cli/descriptions.js`](diffhunk://#diff-4779ca37913f14f5aec15ce530617d866fc03de73d6a32a61264fbc8bdef8a77R15-R16):
Added a description for the `bug` command.

### Additional Changes:
*
[`packages/bananass/src/commands/index.js`](diffhunk://#diff-c96aa353b75dcf3f251899c2993ade50ecacb5517ebbf35188abd138b982e1cdR9):
Imported and exported the `bug` command in the main commands index file.
[[1]](diffhunk://#diff-c96aa353b75dcf3f251899c2993ade50ecacb5517ebbf35188abd138b982e1cdR9)
[[2]](diffhunk://#diff-c96aa353b75dcf3f251899c2993ade50ecacb5517ebbf35188abd138b982e1cdL19-R20)
*
[`packages/bananass/src/commands/bananass-bug/bug.test.js`](diffhunk://#diff-1e9b3d68722f8d1c64c12fce93bc307972f33c547d75b6d804f51e5942e02b0aR1):
Placeholder for future tests for the `bug` command.