feat(bananass): create new `info` command (#167)
This pull request introduces a new `info` command to the `bananass` CLI,
which prints relevant details about the current system for bug
reporting. The key changes include adding the `envinfo` package,
implementing the `info` command, and updating the configuration and
options to support the new command.

### New `info` command implementation:

*
[`packages/bananass/src/cli/bananass-info.js`](diffhunk://#diff-bff245da03f8a2d2d1dc34bee509e9bc4a2eaaa81cf70498e71536284f9267caL9-R27):
Replaced the `warning` description with the new `info` command
implementation, including options and an asynchronous action to load and
log configuration details.
[[1]](diffhunk://#diff-bff245da03f8a2d2d1dc34bee509e9bc4a2eaaa81cf70498e71536284f9267caL9-R27)
[[2]](diffhunk://#diff-bff245da03f8a2d2d1dc34bee509e9bc4a2eaaa81cf70498e71536284f9267caL29-R73)
*
[`packages/bananass/src/commands/bananass-info/index.js`](diffhunk://#diff-e7c4b358eaea1f11779c22944ab318d6cbec3935dcfa0ae4a86853a383ffccfcR1-R3):
Added an export for the new `info` command.
*
[`packages/bananass/src/commands/bananass-info/info.js`](diffhunk://#diff-fff40416892203d981882c1ebf6d060b831d4aa44b16b388686a04c49323dcedR1-R92):
Implemented the `info` command to asynchronously print system details
using `envinfo`.

### Configuration and options updates:

*
[`packages/bananass/src/core/conf/default-config-object/default-config-object.js`](diffhunk://#diff-52a8798a5e892009d2bf6b96aaff12cc552d06b6a08c3338dacfcfb2f15d6b89L69-R71):
Added a default configuration for the `info` command.
*
[`packages/bananass/src/core/cli/options.js`](diffhunk://#diff-dfd07f0b101d852284bb7229d33c6c43fd8ffe2aa88a16484b64079d25550de1L77-R86):
Added a new option `all` for the `info` command to show all information,
including not found items.
*
[`packages/bananass/src/core/structs/config-object-info/config-object-info.js`](diffhunk://#diff-3379af6595a95599682039f1a79cd8ab849d139dd43f5ccaa9b7ca1167d24237L9-R9):
Updated the `ConfigObjectInfo` structure to include the `all` property.
[[1]](diffhunk://#diff-3379af6595a95599682039f1a79cd8ab849d139dd43f5ccaa9b7ca1167d24237L9-R9)
[[2]](diffhunk://#diff-3379af6595a95599682039f1a79cd8ab849d139dd43f5ccaa9b7ca1167d24237L29-R31)

### Additional changes:

*
[`packages/bananass/package.json`](diffhunk://#diff-e74f46a7c449d500fc71c84c894fd8e911cf9fe4e54157b741ff6f9cbf357432R67):
Added the `envinfo` package as a dependency.
*
[`packages/bananass/src/core/constants.js`](diffhunk://#diff-97d6046c52dcaa9f59ecd0bc5a1ed48d3a91d2c8159901b86ba0b4932b532886R81-R90):
Defined `BANANASS_PKG_NAMES` to list relevant package names for the
`info` command.