fix(eslint-config-bananass): update `react/curly-brace-presence` to follow default options (#185)
This pull request includes several changes to the ESLint configuration
files in the `packages/eslint-config-bananass` directory. The changes
primarily focus on updating rule descriptions and modifying rule
settings for better code analysis.

### Updates to rule descriptions:

*
[`packages/eslint-config-bananass/src/rules/import/import-static-analysis.js`](diffhunk://#diff-a415fbe05c5cbf8feb435325f7f58bcced3a908ae247260a9b82300b1f1457ffR117):
Added a link to an issue in the rule description for better context.
*
[`packages/eslint-config-bananass/src/rules/node/node.js`](diffhunk://#diff-c96c91af708b289359a5b68fb41a21628063ede30d509cc8901b9063cacd9309L113-R113):
Updated the description to clarify the use of `require()` expressions.

### Modifications to rule settings:

*
[`packages/eslint-config-bananass/src/rules/react/react.js`](diffhunk://#diff-d2e14b50ea9ff3e279f30d972b03fd16efe19993bd0d6670de8571d2213515c5L199-R199):
Changed the setting for `propElementValues` from 'never' to 'ignore' to
adjust the rule's behavior.