fix(bananass): update `Solution` and `Testcase` types to allow optional input parameters (#220)
This pull request includes changes to the `packages/bananass` module to
refine the structure and typing of solutions and test cases. The most
important changes include modifying the `Solution` function validation,
updating the `Testcase` structure, and reorganizing type definitions.

Refinements to structure and typing:

*
[`packages/bananass/src/core/structs/solution/solution.js`](diffhunk://#diff-1a3d3e2df9dacfb36c9df0d999baf3a817992c8dce1e5b0f5c753b174312f1aeL30-R32):
Modified the `Solution` function validation to accept functions with
zero or one parameter.
*
[`packages/bananass/src/core/structs/testcase/testcase.js`](diffhunk://#diff-2fae1502d373d29acefcff8a8a39a93c2cc5ca1950b90df60ac300f7c2c117deL9-R9):
Added `optional` to the `superstruct` import and updated the `Testcase`
structure to make the `input` property optional.
[[1]](diffhunk://#diff-2fae1502d373d29acefcff8a8a39a93c2cc5ca1950b90df60ac300f7c2c117deL9-R9)
[[2]](diffhunk://#diff-2fae1502d373d29acefcff8a8a39a93c2cc5ca1950b90df60ac300f7c2c117deL30-R30)

Reorganization of type definitions:

*
[`packages/bananass/src/core/types.js`](diffhunk://#diff-0b5f63f76dd8d7069ef461522345f424e381b4ccd8fd1cc3b8683bb57a1f05cdL127-R135):
Updated the `Solution` callback type definition to make the `input`
parameter optional.
*
[`packages/bananass/src/core/types.js`](diffhunk://#diff-0b5f63f76dd8d7069ef461522345f424e381b4ccd8fd1cc3b8683bb57a1f05cdR149-R154):
Moved the `SolutionWithTestcases` type definition to a new location
within the file for better organization.