chore(websites-vitepress): add `@codecov/vite-plugin` and update related configs for bundle analysis (#259)
This pull request introduces a new Codecov Vite plugin to the Vitepress
configuration and updates the dependencies accordingly. The most
important changes include the addition of the Codecov Vite plugin and
the `is-interactive` package, as well as the necessary modifications to
the Vitepress configuration to incorporate these new tools.

### Integration of Codecov Vite Plugin:

*
[`websites/vitepress/.vitepress/config/shared.js`](diffhunk://#diff-73876f6a978dec27dfeeaa8c70437d5664ae37af70af0cda625a8d54373d237bR16-R17):
Added imports for the Codecov Vite plugin and `is-interactive`, and
updated the Vite configuration to include the Codecov Vite plugin with
specific settings, ensuring it only runs in CI environments.
[[1]](diffhunk://#diff-73876f6a978dec27dfeeaa8c70437d5664ae37af70af0cda625a8d54373d237bR16-R17)
[[2]](diffhunk://#diff-73876f6a978dec27dfeeaa8c70437d5664ae37af70af0cda625a8d54373d237bL102-R113)

### Dependency Updates:

*
[`websites/vitepress/package.json`](diffhunk://#diff-10ef64608fffd18cd9a3d2a3b2adb3361c324abf4a0040a210720ab259c6acd2R12-R15):
Added `@codecov/vite-plugin` and `is-interactive` to the
`devDependencies`.